Enable multiple section wrappers
